Fix MultiNest Cube dimension error
When initialised, Cube had a dimension of nPar = TOTAL_PARAM_AMT,
which implicitly contains values associated with fixed parameters.
This dependency was removed as fixed parameters are known at compile
time and hence do not need to be stored dynamically.

Values used in multinest_loglike() for the size of arrays have been
changed to use the function's input parameter sizes. This should
prevent any erroneous behaviour later on.