Add support for HyRec initialisation
HyRec initialisation occurs once before MultiNest is run, and then has
its values passed into a ClassEngine member variable, which then uses it
throughout its lifetime to calculate cosmologies using HyRec as the
recombination code.

This sweeping change is only compatible with the appropriate version of
CLASS containing the external HyRec support. As of writing, this version
may be found in the `phoenix` branch of CLASS as forked by hdp1213.

Also add both CLASS C++ wrappers to the repo. This way, there does not
need to be an implicit reliance on the C++ wrappers shipped with CLASS,
which I found to be troublesome in the long term when more non-standard
parts are bolted on.

Also comment out `sigma8` from the TTTEEE+lowP_pbh.h params file.
Compilation for some reason complains that `sigma8` has been defined
twice, yet I cannot see how this is possible. Definitely a future issue
to tackle!

Also change LCDM parameters in test_multinest to their best-fit 2017
Planck values.

Also add some deallocation in test_multinest to check that memory does
not leak as it is passed through Valgrind.