Prepare pc_multinest for 2D Clark runs
This has entailed adding a new set of parameters, and using the lite
high-l likelihood as opposed to the full likelihood.

There are memory leaks in this commit that need to be cleaned up before
any runs are initiated. Large leaking occurs in HyRec, where for
whatever reason failures occur without deallocating any memory. Should
be able to fix this using a similar technique used in CLASS.