Change NCDM parameters to remove leaking
This parameter change should replicate the physics without having to run
the infamous background_init_ncdm(), which seems to leak memory and I
cannot figure out why.

This commit rounds out the final changes on the pc_multinest side of
things. It is now in alignment with the first 2D Clark run, run 01.