Change parameters for PBH mass mean/width 2D runs
Parameters used now correspond to fixed LCDM and A_planck, fixed PBH
fraction to 1.0, and PBH mass mean and width allowed to vary. Both mean
and width are in log10 space, to effectively explore the lowest areas of
this parameter space. Note for width of 1e-10, the distribution is very
similar to a delta distribution.

Corresponds to pc_multinest_mpi_pbh_dist01 run submitted on 19/12/17,
with no extra fixes to HyRec's high-Tm behaviour.