Update likelihood test
The likelihood test is still broken, I am aware of this. This is because
the current Parameter.hpp file requires both a PBH mass and PBH width,
and no A_planck. At the time this likelihood test was written, the
values that were required were just the vanilla LCDM + A_planck
parameters. I'll fix this eventually, but for now expect it to be broken
and (although run) not given any intelligent result.

To clarify, this commit reflects updates to the other tests in that it
should now compile and run with no complaints or seg faults (provided
the `phoenix` branch of CLASS is used, of course!), but due to what was
mentioned it gives garbage output.