Create apps folder
This was done so when I want to add more binaries that aren't
necessarily tests, I don't have to worry about accidentally including
object files of other programs and hitting my head on the multiple mains
issue. The apps/ directory is now for programs with a main() function
that aren't doing any testing.