Fix gaussian prior calculation bug
When parameters with gaussian priors are declared as fixed in the
parameter enumeration, the program erroneously drew the parameter values
from the input parameter array, instead of from the global g_value
array. This lead to very large underestimations of the log likelihood,
where in the case of a fixed A_planck, values of H0 were being taken
from the input array instead. This lead to a gaussian correction of the
order of -3e8, swamping any actual changes in the likelihood.

Now, the program checks in a manner similar to ClikObject to determine
which array to use in this calculation. Another point to make is that
this check does not exist in pc_multinest, which must be due to the
different way it handles the input array of values.