Fix another gaussian likelihood bug
This only appeared in the unlikely case that all parameters are fixed
and the full high-l likelihood is being used, such as when you are
performing likelihood profiling. This was fixed by a more liberal
application of an inline function that does the heavy lifting for you,
so you don't need to sweat about where the actual value you want will
be.