Update profiler best-fit values
These come from base_planck_14, and are currently unchecked, so expect
some news soon about what's actually happening.

To be more explicit, these are only for the lite high-l likelihood, so
all of the fixed values for the full likelihood still have their old
values and so will not work properly anymore!