Add tau/H0 degeneracy check runs
The new set of parameters fixes all LCDM parameters and gives the option
for the user to allow one of these to vary. It is currently set up to
vary 100xtheta_s, (which is effectively H0), but also works with
tau_reio and other parameters with the appropriate uniform priors.

This work is quite old at this point, and is being committed to allow
further work on the program to proceed as required.